+++ src/Test/Unit/Model/Config/Source/LogLevelTest.php
@@ -1,12 +1,17 @@
 <?php
-
+/**
+ * Copyright (c) 2024 Mohamed EL Mrabet
+ * CleatSquad - https://cleatsquad.dev
+ *
+ * This file is part of the CleatSquad_LogStream module.
+ * Licensed under the MIT License. See the LICENSE file in the module root.
+ */
 declare(strict_types=1);
 
 namespace CleatSquad\LogStream\Test\Unit\Model\Config\Source;
 
 use PHPUnit\Framework\TestCase;
 use CleatSquad\LogStream\Model\Config\Source\LogLevel;
-use Monolog\Logger;
 
 /**
  * Class LogLevelTest
@@ -16,21 +21,24 @@ class LogLevelTest extends TestCase
 {
     /**
      * Test the toOptionArray method.
+     *
+     * Uses fixed integer values that are the same in both Monolog 2.x and 3.x
      */
     public function testToOptionArray(): void
     {
         $logLevel = new LogLevel();
         $options = $logLevel->toOptionArray();
 
+        // These integer values are the same in both Monolog 2.x and 3.x
         $expectedOptions = [
-            ['value' => Logger::DEBUG, 'label' => 'DEBUG'],
-            ['value' => Logger::INFO, 'label' => 'INFO'],
-            ['value' => Logger::NOTICE, 'label' => 'NOTICE'],
-            ['value' => Logger::WARNING, 'label' => 'WARNING'],
-            ['value' => Logger::ERROR, 'label' => 'ERROR'],
-            ['value' => Logger::CRITICAL, 'label' => 'CRITICAL'],
-            ['value' => Logger::ALERT, 'label' => 'ALERT'],
-            ['value' => Logger::EMERGENCY, 'label' => 'EMERGENCY'],
+            ['value' => 100, 'label' => 'DEBUG'],
+            ['value' => 200, 'label' => 'INFO'],
+            ['value' => 250, 'label' => 'NOTICE'],
+            ['value' => 300, 'label' => 'WARNING'],
+            ['value' => 400, 'label' => 'ERROR'],
+            ['value' => 500, 'label' => 'CRITICAL'],
+            ['value' => 550, 'label' => 'ALERT'],
+            ['value' => 600, 'label' => 'EMERGENCY'],
         ];
 
         $this->assertEquals($expectedOptions, $options);